requirements.txt
aiohttp==3.11.10
beautifulsoup4==4.12.3
PyYAML==6.0.2